chore: move process-related methods to its file
